renamed storage, now image
