+++ backend/src/app.module.ts
@@ -18,7 +18,7 @@ import {Te} from './entities/products/product-te.entity'
 import { Chocolate } from './entities/products/product-chocolate.entity';
 import { Accesorio } from './entities/products/product-accesorio.entity';
 import { OrderModule } from './modules/order/order.module';
-import { ImageController } from './modules/storage/image.controller';
+import { ImageController } from './modules/images/image.controller';
 import { MercadoPagoModule } from './modules/mercadopago/mercado-pago.module';
 import { JwtModule } from '@nestjs/jwt';
 import { TestimonyModule } from './modules/testimony/testimony.module';
@@ -26,8 +26,8 @@ import { Testimony } from './entities/testimony.entity';
 import { CategoryModule } from './modules/categories/category.module';
 import { StorageOrderModule } from './modules/storageOrder/storage-order.module';
 import { Storage } from './entities/storage.entity';
-import { ImageService } from './modules/storage/image.service';
-import { ImageModule } from './modules/storage/image.module';
+import { ImageService } from './modules/images/image.service';
+import { ImageModule } from './modules/images/image.module';
 
 @Module({
   imports: [
