+++ backend/src/modules/images/image.controller.ts
