+++ backend/src/modules/images/image.service.ts
