+++ backend/src/modules/products/products.module.ts
@@ -6,7 +6,7 @@ import { Product } from 'src/entities/products/product.entity';
 import { Category } from 'src/entities/category.entity';
 import { Coffee } from 'src/entities/products/product-coffee.entity';
 import { Subproduct } from 'src/entities/products/subprodcut.entity';
-import { ImageModule } from '../storage/image.module';
+import { ImageModule } from '../images/image.module';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Product,Category,Coffee, Subproduct]), ImageModule],
